% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{as.character,tinytable-method}
\alias{as.character,tinytable-method}
\title{Convert a tinytable S4 object to a string}
\usage{
\S4method{as.character}{tinytable}(x)
}
\arguments{
\item{x}{A data frame, data table, or tibble to be rendered as a table.}
}
\description{
Convert a tinytable S4 object to a string
}
\keyword{internal}
