% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapOptions.R
\name{bootstrapOptions}
\alias{bootstrapOptions}
\title{HTML Bootstrap Options}
\usage{
bootstrapOptions(class = "table", css = "", ...)
}
\arguments{
\item{class}{TODO}

\item{css}{TODO}
}
\description{
HTML Bootstrap Options
}
