% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{colnames,tinytable-method}
\alias{colnames,tinytable-method}
\title{Method for a tinytable S4 object}
\usage{
\S4method{colnames}{tinytable}(x)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}
}
\description{
Method for a tinytable S4 object
}
\keyword{internal}
