% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{df_to_json}
\alias{df_to_json}
\title{Convert data frame or list to JSON}
\usage{
df_to_json(data)
}
\arguments{
\item{data}{A data frame or list to convert to JSON}
}
\value{
JSON string
}
\description{
Simple helper function to convert a data frame or list to JSON format
}
\keyword{internal}
