% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{dim,tinytable-method}
\alias{dim,tinytable-method}
\title{Dimensions a tinytable S4 object}
\usage{
\S4method{dim}{tinytable}(x)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}
}
\description{
Dimensions a tinytable S4 object
}
\keyword{internal}
