% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_grid_dataframe.R
\name{finalize,tinytable_dataframe-method}
\alias{finalize,tinytable_dataframe-method}
\title{tinytable S4 method}
\usage{
\S4method{finalize}{tinytable_dataframe}(x)
}
\description{
tinytable S4 method
}
\keyword{internal}
