% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{finalize}
\alias{finalize}
\title{Apply final settings to a tinytable}
\usage{
finalize(x, ...)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}

\item{...}{Additional arguments are ignored}
}
\description{
Apply final settings to a tinytable
}
\keyword{internal}
