% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_bootstrap.R
\name{group_eval,tinytable_bootstrap-method}
\alias{group_eval,tinytable_bootstrap-method}
\title{tinytable S4 method}
\usage{
\S4method{group_eval}{tinytable_bootstrap}(x, i = NULL, j = NULL, indent = 1, ...)
}
\description{
tinytable S4 method
}
\keyword{internal}
