% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_grid_dataframe.R
\name{group_eval,tinytable_dataframe-method}
\alias{group_eval,tinytable_dataframe-method}
\title{tinytable S4 method}
\usage{
\S4method{group_eval}{tinytable_dataframe}(x)
}
\description{
tinytable S4 method
}
\keyword{internal}
