% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tabularray.R
\name{group_eval,tinytable_tabularray-method}
\alias{group_eval,tinytable_tabularray-method}
\title{tinytable S4 method}
\usage{
\S4method{group_eval}{tinytable_tabularray}(x, i = NULL, j = NULL, indent = 1, ...)
}
\description{
tinytable S4 method
}
\keyword{internal}
