% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_group.R
\name{group_eval_j,tinytable_html-method}
\alias{group_eval_j,tinytable_html-method}
\title{tinytable S4 method}
\usage{
\S4method{group_eval_j}{tinytable_html}(x, i = NULL, j = NULL, ihead = NULL, ...)
}
\description{
tinytable S4 method
}
\keyword{internal}
