% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tabulator.R
\name{group_eval_j,tinytable_tabulator-method}
\alias{group_eval_j,tinytable_tabulator-method}
\title{tinytable S4 method}
\usage{
\S4method{group_eval_j}{tinytable_tabulator}(x, i = NULL, j = NULL, ...)
}
\description{
tinytable S4 method
}
\keyword{internal}
