% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst_group.R
\name{group_eval_j,tinytable_typst-method}
\alias{group_eval_j,tinytable_typst-method}
\title{tinytable S4 method}
\usage{
\S4method{group_eval_j}{tinytable_typst}(x, i = NULL, j = NULL, indent = 0, ...)
}
\description{
tinytable S4 method
}
\keyword{internal}
