% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tt.R
\name{group_tt}
\alias{group_tt}
\title{Spanning labels to identify groups of rows or columns}
\usage{
group_tt(x, i, j, italic = TRUE, indent = 1)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}

\item{i}{Row index or indices where the styling should be applied. Can be a single value or a vector. If \code{colspan} is used, \code{i} must be of length 1.}

\item{j}{Column index or indices where the styling should be applied. Can be a single value or a vector. If \code{colspan} is used, \code{j} must be of length 1.}

\item{italic}{Logical; if \code{TRUE}, text is styled in italic.}

\item{indent}{integer number of \code{pt} to use when indenting the non-labelled rows.}
}
\description{
Spanning labels to identify groups of rows or columns
}
