% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tt.R
\name{group_tt_cols_by_delim}
\alias{group_tt_cols_by_delim}
\title{Group columns by a delimiter}
\usage{
group_tt_cols_by_delim(
  x,
  j = NULL,
  delim = "_",
  split = c("last", "first"),
  ...
)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}

\item{j}{An integer vector of column indices to group. If NULL, uses all columns with the delim.}

\item{delim}{A string that is used as a delimiter to identify groups of columns. The columns are grouped by the part of the column names before the delimiter. Currently only allows splitting at a single location (one layer of grouping).}

\item{split}{A string that specifies how to handle multiple delimiters in the column names. If \code{"last"} (default), the group and sub-header are split at the last occuring delim. If \code{"first"}, splitting occurs at the first occurrence.}

\item{...}{Other arguments are ignored.}
}
\value{
An object of class \code{tt} representing the table.
}
\description{
Group columns by a delimiter
}
\details{
Warning: The \code{style_tt()} can normally be used to style the group headers, as expected, but that feature is not available for Markdown and Word tables.
}
\section{Word and Markdown limitations}{


Markdown and Word tables only support these styles: italic, bold, strikeout. The \code{width} argument is also unavailable
Moreover, the \code{style_tt()} function cannot be used to style headers inserted by the \code{group_tt()} function;
instead, you should style the headers directly in the header definition using markdown syntax:
\code{group_tt(i = list("*italic header*" = 2))}. These limitations are due to the fact that there is no markdown
syntax for the other options, and that we create Word documents by converting a markdown table to .docx
via the Pandoc software.
}

\examples{
# Only columns with a delimiter are touched:
data.frame(id = 1, a_b = 2, a_c = 3, b_b = 4, b_c = 5, e = 6) |>
  tt() |>
  group_tt_cols_by_delim()
}
