% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{initialize,tinytable-method}
\alias{initialize,tinytable-method}
\title{Method for a tinytable S4 object}
\usage{
\S4method{initialize}{tinytable}(.Object, ...)
}
\arguments{
\item{...}{Additional arguments are ignored}
}
\description{
Method for a tinytable S4 object
}
\keyword{internal}
