% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.tinytable}
\alias{knit_print.tinytable}
\title{Print a tinytable object in knitr}
\usage{
\method{knit_print}{tinytable}(
  x,
  output = get_option("tinytable_print_output", default = NULL),
  ...
)
}
\value{
A string with class \code{knit_asis} to be printed in Rmarkdown or Quarto documents.
}
\description{
Print a tinytable object in knitr
}
\keyword{internal}
