% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.tinytable_bootstrap}
\alias{knit_print.tinytable_bootstrap}
\title{Print a tinytable object in knitr}
\usage{
knit_print.tinytable_bootstrap(x, ...)
}
\description{
Print a tinytable object in knitr
}
\keyword{internal}
