% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexOptions.R
\name{latexOptions}
\alias{latexOptions}
\title{LaTeX options}
\usage{
latexOptions(
  environment = getOption("tt_latexOptions_environment", default = "table+tblr"),
  extendable = getOption("tt_latexOptions_extendable", default = FALSE),
  placement = getOption("tt_latexOptions_placement", default = NULL),
  theme = getOption("tt_latexOptions_theme", default = "booktabs"),
  ...
)
}
\description{
LaTeX options
}
\section{LaTeX \code{tabularray} options}{

\subsection{Table: \code{tt()}}{

Inner specifications:\tabular{lll}{
   Key \tab Description and Values \tab Initial Value \cr
   \code{rulesep} \tab space between two hlines or vlines \tab \verb{2pt} \cr
   \code{stretch} \tab stretch ratio for struts added to cell text \tab \code{1} \cr
   \code{abovesep} \tab set vertical space above every row \tab \verb{2pt} \cr
   \code{belowsep} \tab set vertical space below every row \tab \verb{2pt} \cr
   \code{rowsep} \tab set vertical space above and below every row \tab \verb{2pt} \cr
   \code{leftsep} \tab set horizontal space to the left of every column \tab \verb{6pt} \cr
   \code{rightsep} \tab set horizontal space to the right of every column \tab \verb{6pt} \cr
   \code{colsep} \tab set horizontal space to both sides of every column \tab \verb{6pt} \cr
   \code{hspan} \tab horizontal span algorithm: \code{default}, \code{even}, or \code{minimal} \tab \code{default} \cr
   \code{vspan} \tab vertical span algorithm: \code{default} or \code{even} \tab \code{default} \cr
   \code{baseline} \tab set the baseline of the table \tab \code{m} \cr
}


Outer specifications:\tabular{lll}{
   Key \tab Description and Values \tab Initial Value \cr
   \code{baseline} \tab set the baseline of the table \tab \code{m} \cr
   \code{long} \tab change the table to a long table \tab None \cr
   \code{tall} \tab change the table to a tall table \tab None \cr
   \code{expand} \tab you need this key to use verb commands \tab None \cr
}

}

\subsection{Cells: \code{style_cell()}}{\tabular{lll}{
   Key \tab Description and Values \tab Initial Value \cr
   \code{halign} \tab horizontal alignment: \code{l} (left), \code{c} (center), \code{r} (right) or \code{j} (justify) \tab \code{j} \cr
   \code{valign} \tab vertical alignment: \code{t} (top), \code{m} (middle), \code{b} (bottom), \code{h} (head) or \code{f} (foot) \tab \code{t} \cr
   \code{wd} \tab width dimension \tab None \cr
   \code{bg} \tab background color name \tab None \cr
   \code{fg} \tab foreground color name \tab None \cr
   \code{font} \tab font commands \tab None \cr
   \code{mode} \tab set cell mode: \code{math}, \code{imath}, \code{dmath} or \code{text} \tab None \cr
   \code{cmd} \tab execute command for the cell text \tab None \cr
   \code{preto} \tab prepend text to the cell \tab None \cr
   \code{appto} \tab append text to the cell \tab None \cr
   \code{r} \tab number of rows the cell spans \tab 1 \cr
   \code{c} \tab number of columns the cell spans \tab 1 \cr
}

}

\subsection{Rows: \code{style_row()}}{\tabular{lll}{
   Key \tab Description and Values \tab Initial Value \cr
   \code{halign} \tab horizontal alignment: \code{l} (left), \code{c} (center), \code{r} (right) or \code{j} (justify) \tab \code{j} \cr
   \code{valign} \tab vertical alignment: \code{t} (top), \code{m} (middle), \code{b} (bottom), \code{h} (head) or \code{f} (foot) \tab \code{t} \cr
   \code{ht} \tab height dimension \tab None \cr
   \code{bg} \tab background color name \tab None \cr
   \code{fg} \tab foreground color name \tab None \cr
   \code{font} \tab font commands \tab None \cr
   \code{mode} \tab set mode for row cells: \code{math}, \code{imath}, \code{dmath} or \code{text} \tab None \cr
   \code{cmd} \tab execute command for every cell text \tab None \cr
   \code{abovesep} \tab set vertical space above the row \tab \verb{2pt} \cr
   \code{belowsep} \tab set vertical space below the row \tab \verb{2pt} \cr
   \code{rowsep} \tab set vertical space above and below the row \tab \verb{2pt} \cr
   \code{preto} \tab prepend text to every cell (like \code{>} specifier in \code{rowspec}) \tab None \cr
   \code{appto} \tab append text to every cell (like \code{<} specifier in \code{rowspec}) \tab None \cr
}

}

\subsection{Columns: \code{style_column()}}{\tabular{lll}{
   Key \tab Description and Values \tab Initial Value \cr
   \code{halign} \tab horizontal alignment: \code{l} (left), \code{c} (center), \code{r} (right) or \code{j} (justify) \tab \code{j} \cr
   \code{valign} \tab vertical alignment: \code{t} (top), \code{m} (middle), \code{b} (bottom), \code{h} (head) or \code{f} (foot) \tab \code{t} \cr
   \code{wd} \tab width dimension \tab None \cr
   \code{co} \tab coefficient for the extendable column (\code{X} column) \tab None \cr
   \code{bg} \tab background color name \tab None \cr
   \code{fg} \tab foreground color name \tab None \cr
   \code{font} \tab font commands \tab None \cr
   \code{mode} \tab set mode for column cells: \code{math}, \code{imath}, \code{dmath} or \code{text} \tab None \cr
   \code{cmd} \tab execute command for every cell text \tab None \cr
   \code{leftsep} \tab set horizontal space to the left of the column \tab \verb{6pt} \cr
   \code{rightsep} \tab set horizontal space to the right of the column \tab \verb{6pt} \cr
   \code{colsep} \tab set horizontal space to both sides of the column \tab \verb{6pt} \cr
   \code{preto} \tab prepend text to every cell (like \code{>} specifier in \code{colspec}) \tab None \cr
   \code{appto} \tab append text to every cell (like \code{<} specifier in \code{colspec}) \tab None \cr
}

}

\subsection{hlines}{\tabular{lll}{
   Key \tab Description and Values \tab Initial Value \cr
   \code{dash} \tab dash style: \code{solid}, \code{dashed} or \code{dotted} \tab \code{solid} \cr
   \code{text} \tab replace hline with text (like \code{!} specifier in \code{rowspec}) \tab None \cr
   \code{wd} \tab rule width dimension \tab \verb{0.4pt} \cr
   \code{fg} \tab rule color name \tab None \cr
   \code{leftpos} \tab crossing or trimming position at the left side \tab \code{1} \cr
   \code{rightpos} \tab crossing or trimming position at the right side \tab \code{1} \cr
   \code{endpos} \tab adjust leftpos/rightpos for only the leftmost/rightmost column \tab \code{false} \cr
}

}

\subsection{vlines}{\tabular{lll}{
   Key \tab Description and Values \tab Initial Value \cr
   \code{dash} \tab dash style: \code{solid}, \code{dashed} or \code{dotted} \tab \code{solid} \cr
   \code{text} \tab replace vline with text (like \code{!} specifier in \code{colspec}) \tab None \cr
   \code{wd} \tab rule width dimension \tab \verb{0.4pt} \cr
   \code{fg} \tab rule color name \tab None \cr
   \code{abovepos} \tab crossing or trimming position at the above side \tab \code{0} \cr
   \code{belowpos} \tab crossing or trimming position at the below side \tab \code{0} \cr
}

}
}

