% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{names,tinytable-method}
\alias{names,tinytable-method}
\title{Column names of a tinytable}
\usage{
\S4method{names}{tinytable}(x)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}
}
\description{
Column names of a tinytable
}
\keyword{internal}
