% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{ncol,tinytable-method}
\alias{ncol,tinytable-method}
\title{Method for a tinytable S4 object}
\usage{
\S4method{ncol}{tinytable}(x)
}
\arguments{
\item{x}{A data frame, data table, or tibble to be rendered as a table.}
}
\description{
Method for a tinytable S4 object
}
\keyword{internal}
