% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_bootstrap.R
\name{options_bootstrap}
\alias{options_bootstrap}
\title{HTML Bootstrap Options}
\usage{
options_bootstrap(class = "table", css = "", ...)
}
\arguments{
\item{class}{TODO}

\item{css}{TODO}
}
\description{
HTML Bootstrap Options
}
