% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{pkgdown_print.tinytable}
\alias{pkgdown_print.tinytable}
\title{Print a tinytable object in pkgdown}
\usage{
pkgdown_print.tinytable(x, visible = TRUE)
}
\arguments{
\item{x}{A tinytable object}

\item{visible}{Whether the object is visible}
}
\value{
HTML output for pkgdown examples
}
\description{
Print a tinytable object in pkgdown
}
\keyword{internal}
