% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind.R
\name{rbind2,tinytable,ANY-method}
\alias{rbind2,tinytable,ANY-method}
\title{Combine \code{tinytable} objects by rows (vertically)}
\usage{
\S4method{rbind2}{tinytable,ANY}(x, y, fill = TRUE, use.names = TRUE, headers = FALSE, ...)
}
\arguments{
\item{fill}{\code{TRUE} fills missing columns with NAs. By default \code{FALSE}.}

\item{use.names}{\code{TRUE} binds by matching column name, \code{FALSE} by position. `check` (default) warns if all items don't have the same names in the same order and then currently proceeds as if `use.names=FALSE` for backwards compatibility (\code{TRUE} in future); see news for v1.12.2.}

\item{headers}{Logical. TRUE inserts the colnames of \code{y} as an extra row between the two tables.}
}
\description{
Two \code{tinytable}x
objects are combined vertically.

Any \code{format_tt()} calls applied to \code{x} or \code{y} will be evaluated before binding, to allow distinct formatting for each panel.

Calls to other \code{tinytable} function such as \code{style_tt()} or \code{group_tt()} are ignored when applied to \code{x} or \code{y}. These functions should be applied to the final table instead.
}
