% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind2.R
\name{rbind2}
\alias{rbind2}
\alias{rbind2,tinytable}
\title{Combine \code{tinytable} objects by rows (vertically)}
\usage{
\S4method{rbind2}{tinytable,ANY}(x, y, fill = TRUE, use.names = TRUE, headers = TRUE, ...)
}
\arguments{
\item{x}{\code{tinytable} object}

\item{y}{\code{tinytable} object}

\item{fill}{Logical. TRUE fills the missing columns with \code{NA}}

\item{use.names}{‘TRUE’ binds by matching column name, ‘FALSE’ by position}

\item{headers}{Logical. TRUE inserts the colnames of \code{y} as an extra row between the two tables.}
}
\description{
Two \code{tinytable}x
objects are combined vertically.

Any \code{format_tt()} calls applied to \code{x} or \code{y} will be evaluated before binding, to allow distinct formatting for each panel.

Calls to other \code{tinytable} function such as \code{style_tt()} or \code{group_tt()} are ignored when applied to \code{x} or \code{y}. These functions should be applied to the final table instead.

This function requires the \code{data.table} package.
}
