% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_grid.R
\name{style_eval,tinytable_grid-method}
\alias{style_eval,tinytable_grid-method}
\title{tinytable S4 method}
\usage{
\S4method{style_eval}{tinytable_grid}(
  x,
  i = NULL,
  j = NULL,
  bold = FALSE,
  italic = FALSE,
  monospace = FALSE,
  underline = FALSE,
  strikeout = FALSE,
  rowspan = NULL,
  colspan = NULL,
  ...
)
}
\description{
tinytable S4 method
}
\keyword{internal}
