% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{style_eval}
\alias{style_eval}
\title{Apply style settings to a tinytable}
\usage{
style_eval(x, ...)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}

\item{...}{Additional arguments are ignored}
}
\description{
Apply style settings to a tinytable
}
\keyword{internal}
