% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_grid.R
\name{style_grid_internal}
\alias{style_grid_internal}
\title{tinytable S4 method}
\usage{
style_grid_internal(
  x,
  i = NULL,
  j = NULL,
  bold = FALSE,
  italic = FALSE,
  monospace = FALSE,
  underline = FALSE,
  strikeout = FALSE,
  rowspan = NULL,
  colspan = NULL,
  ...
)
}
\description{
tinytable S4 method
}
\keyword{internal}
