% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{style_row}
\alias{style_row}
\title{Style Rows in a \code{tinytable}}
\usage{
style_row(
  x,
  i = NULL,
  color = NULL,
  background = NULL,
  bold = FALSE,
  italic = FALSE,
  latex = latexOptions(),
  html = htmlOptions()
)
}
\arguments{
\item{x}{A table created by the \code{tt()} function.}

\item{i}{Integer vector indicating row positions. If \code{NULL}, applies to all rows.}

\item{color}{String specifying the text color. If \code{NULL}, no color is applied.}

\item{background}{String specifying the background color. If \code{NULL}, no background is applied.}

\item{bold}{Logical; if \code{TRUE}, applies bold styling to the text.}

\item{italic}{Logical; if \code{TRUE}, applies italic styling to the text.}

\item{latex}{sytling options for the \code{tabularray} framework for LaTeX. See \code{?latexOptions} for details.}

\item{html}{styling options for the Bootstrap framework for HTML. See \code{?htmlOptions} for details.}
}
\value{
Returns the modified table object with the specified cell styles applied.
}
\description{
This function applies styling to specified parts of a table created by the \code{tt()} function of the \code{tinytable} package.

The arguments described below allow users to customize text style and color. The \code{latex} and \code{html} arguments allow deep customization using the \code{tabularray} package for LaTeX, and the Bootstrap framework for HTML. For details, see:
\itemize{
\item \code{?latexOptions}
\item \code{?htmlOptions}
}
}
