% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.tinytable}
\alias{subset.tinytable}
\title{Subsetting a \code{tinytable} object}
\usage{
\method{subset}{tinytable}(x, subset, select, drop = FALSE, ...)
}
\arguments{
\item{x}{object to be subsetted.}

\item{subset}{logical expression indicating elements or rows to keep:
    missing values are taken as false.}

\item{select}{expression, indicating columns to select from a
    data frame.}

\item{drop}{passed on to \code{[} indexing operator.}

\item{...}{further arguments to be passed to or from other methods.}
}
\description{
Return subsets \code{tinytable} which meet conditions.
}
