% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bootstrap.R
\name{theme_bootstrap}
\alias{theme_bootstrap}
\title{HTML-specific styles and options}
\usage{
theme_bootstrap(
  x,
  i = NULL,
  j = NULL,
  class = NULL,
  css = NULL,
  css_rule = NULL,
  ...
)
}
\arguments{
\item{x}{A table object created by \code{tt()}.}

\item{i}{Numeric vector, logical matrix, or string.
\itemize{
\item Numeric vector: Row indices where the styling should be applied. Can be a single value or a vector.
\item Logical matrix: A matrix with the same number of rows and columns as \code{x}. \code{i=0} is the header, and negative values are higher level headers. Row indices refer to rows \emph{after} the insertion of row labels by \code{group_tt()}, when applicable.
\item String: Table components "caption", "colnames", "groupi" (row group labels), "~groupi" (non-group rows), "groupj" (column group labels), "notes".
}}

\item{j}{Column indices where the styling should be applied. Can be:
\itemize{
\item Integer vectors indicating column positions.
\item Character vector indicating column names.
\item A single string specifying a Perl-style regular expression used to match column names.
}}

\item{class}{String. Bootstrap table class such as \code{"table"}, \code{"table table-dark"} or \code{"table table-dark table-hover"}. See the bootstrap documentation.}

\item{css}{Character vector. CSS style declarations to be applied to every cell defined by \code{i} and \code{j} (ex: \code{"font-weight: bold"}).}

\item{css_rule}{String. Complete CSS rules (with curly braces, semicolon, etc.) that apply to the table class specified by the \code{bootstrap_class} argument.}

\item{...}{extra arguments are ignored}
}
\description{
HTML-specific styles and options
}
