% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_default.R
\name{theme_default}
\alias{theme_default}
\title{Deprecated: Use theme_tinytable() instead}
\usage{
theme_default(x, ...)
}
\arguments{
\item{x}{A tinytable object.}

\item{...}{Additional arguments are ignored.}
}
\value{
A modified \code{tinytable} object.
}
\description{
\code{theme_default()} has been renamed to \code{theme_tinytable()}. This function is
deprecated and will be removed in a future version.
}
