% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grid.R
\name{theme_grid}
\alias{theme_grid}
\title{Grid theme with borders around all cells}
\usage{
theme_grid(x, ...)
}
\arguments{
\item{x}{A \code{tinytable} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A modified \code{tinytable} object.
}
\description{
Grid theme with borders around all cells
}
