% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_html.R
\name{theme_html}
\alias{theme_html}
\title{HTML-specific styles and options}
\usage{
theme_html(x, engine = "bootstrap", ...)
}
\arguments{
\item{x}{A \code{tinytable} object.}

\item{engine}{Character string specifying the HTML engine. If "raw", strips bootstrap styling and returns a simple table.}

\item{...}{Additional arguments.}
}
\description{
HTML-specific styles and options
}
