% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_revealjs.R
\name{theme_revealjs}
\alias{theme_revealjs}
\title{RevealJS presentation theme}
\usage{
theme_revealjs(
  x,
  css = get_option("tinytable_revealjs_css", default = "light"),
  fontsize = get_option("tinytable_revealjs_fontsize", default = 0.8),
  fontsize_caption = get_option("tinytable_revealjs_fontsize_caption", default = 1)
)
}
\arguments{
\item{x}{A \code{tinytable} object.}

\item{css}{String. CSS theme: "light" (default) or "dark".}

\item{fontsize}{Numeric. Font size multiplier for table content.}

\item{fontsize_caption}{Numeric. Font size multiplier for table captions.}
}
\value{
A modified \code{tinytable} object.
}
\description{
RevealJS presentation theme
}
