% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_striped.R
\name{theme_striped}
\alias{theme_striped}
\title{Striped theme with alternating row colors}
\usage{
theme_striped(x, ...)
}
\arguments{
\item{x}{A \code{tinytable} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
A modified \code{tinytable} object.
}
\description{
Striped theme with alternating row colors
}
