% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tinytable.R
\name{theme_tinytable}
\alias{theme_tinytable}
\title{Default theme for TinyTable}
\usage{
theme_tinytable(x, ...)
}
\arguments{
\item{x}{A tinytable object.}

\item{...}{Additional arguments are ignored.}
}
\value{
A modified \code{tinytable} object.
}
\description{
Default theme for TinyTable
}
