% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_void.R
\name{theme_void}
\alias{theme_void}
\title{Theme for a void table}
\usage{
theme_void(x, ...)
}
\arguments{
\item{x}{A tinytable object.}

\item{...}{Additional arguments are ignored.}
}
\description{
This function calls \code{strip_tt()} to remove all the styles, groups, and formatting applied to a \code{tinytable} object. It returns a nearly blank table, with only the cell information. Warning: since this function strips the \code{tinytable} object, the order in which it is called in a pipeline matters.
}
