\name{IttyBittyTable-package}
\alias{IttyBittyTable-package}
\alias{IttyBittyTable}
\docType{package}
\title{\packageTitle{IttyBittyTable}}
\description{\packageDescription{IttyBittyTable}}
\section{Package Content}{\packageIndices{IttyBittyTable}}
\author{\packageAuthor{IttyBittyTable}}
\section{Maintainer}{\packageMaintainer{IttyBittyTable}}
\keyword{package}
