% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tt.R
\name{group_tt}
\alias{group_tt}
\title{Spanning labels to identify groups of rows or columns}
\usage{
group_tt(x, i, j, italic = TRUE, rule = TRUE, indent = 1)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}

\item{rule}{logical TRUE to insert a horizontal rule below group labels}

\item{indent}{integer number of \code{pt} to use when indenting the non-labelled rows.}
}
\description{
Spanning labels to identify groups of rows or columns
}
