% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_tt.R
\name{group_tt}
\alias{group_tt}
\title{Spanning labels to identify groups of rows or columns}
\usage{
group_tt(x, i = NULL, j = NULL, indent = 1, ...)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}

\item{i}{Row indices where the styling should be applied. Can be a single value or a vector. If \code{colspan} is used, \code{i} must be of length 1. When i=0, the header is styled.}

\item{j}{Column indices where the styling should be applied. Can be a single value, a vector, or a Perl-style regular expression applied to column names of the original data frame. If \code{colspan} is used, \code{j} must be of length 1.}

\item{indent}{integer number of \code{pt} to use when indenting the non-labelled rows.}

\item{...}{All additional arguments (ex: \code{italic}, \code{bold}, \code{color}) are automatically passed to the \code{style_tt()} function and applied to the labels.}
}
\description{
Spanning labels to identify groups of rows or columns
}
