% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{initialize,tinytable-method}
\alias{initialize,tinytable-method}
\title{Method for a tinytable S4 object}
\usage{
\S4method{initialize}{tinytable}(
  .Object,
  data = data.frame(),
  table = data.frame(),
  caption = NULL,
  notes = NULL,
  theme = list("default"),
  placement = NULL,
  width = NULL
)
}
\arguments{
\item{caption}{A string that will be used as the caption of the table.}

\item{notes}{Notes to append to the bottom of the table. This argument accepts several different inputs:
\itemize{
\item Single string insert a single note: \code{"blah blah"}
\item Multiple strings insert multiple notes sequentially: \code{list("Hello world", "Foo bar")}
\item A named list inserts a list with the name as superscript: \code{list("a" = list("Hello World"))}
\item A named list with positions inserts markers as superscripts inside table cells: \code{list("a" = list(i = 0:1, j = 2, text = "Hello World"))}
}}

\item{theme}{Function or string.
\itemize{
\item String: grid, resize, multipage, placement, striped, void, bootstrap, tabular
\item Function: Applied to the \code{tinytable} object.
}}

\item{width}{A numeric vector of length 1 or equal to the number of columns in \code{x}. Each entry must be between 0 and 1, and the sum of the \code{width} vector must also be smaller than or equal to 1. If a single numeric value is provided, \code{width} represents the width of the full table in proportion of the line width. If a vector is supplied, each value represents the horizontal proportion taken up by each column.}
}
\description{
Method for a tinytable S4 object
}
\keyword{internal}
