% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.tinytable}
\alias{knit_print.tinytable}
\title{Print a tinytable object in knitr}
\usage{
knit_print.tinytable(
  x,
  output = getOption("tinytable_print_output", default = NULL),
  ...
)
}
\description{
Print a tinytable object in knitr
}
\keyword{internal}
