% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_class.R
\name{names,tinytable-method}
\alias{names,tinytable-method}
\title{Method for a tinytable S4 object}
\usage{
\S4method{names}{tinytable}(x)

\S4method{names}{tinytable}(x)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}
}
\description{
Method for a tinytable S4 object

Column names of a tinytable
}
\keyword{internal}
