% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tt.R
\name{plot_tt}
\alias{plot_tt}
\title{Insert images and inline plots into tinytable objects}
\usage{
plot_tt(
  x,
  i = NULL,
  j = NULL,
  height = 1,
  plot_asp = 1/3,
  plot_data = NULL,
  plot_fun = NULL,
  path_img = NULL,
  path_assets = "tinytable_assets"
)
}
\arguments{
\item{x}{A tinytable object.}

\item{i}{Integer vector, the row indices where images are to be inserted. If \code{NULL},
images will be inserted in all rows.}

\item{j}{Integer vector, the column indices where images are to be inserted. If \code{NULL},
images will be inserted in all columns.}

\item{height}{Numeric, the height of the images in the table. Default is 2.}

\item{plot_data}{Optional, a list of data frames to be used with a custom plotting function.}

\item{plot_fun}{Optional, a list of functions to generate plots from the data in \code{plot_data}. Valid functions include:
\itemize{
\item Functions that return \code{ggplot2} objects.
\item Functions that return another function which generates a base \code{R} plot, ex: \code{function(x) {function() hist(x)}}
\item See the tutorial on the \code{tinytable} website for more information.
}}

\item{path_img}{Character vector, the paths to the images to be inserted. Paths are relative to the main table file or Quarto (Rmarkdown) document.}

\item{path_assets}{Character, the directory to store generated assets.}
}
\value{
A modified tinytable object with images or plots inserted.
}
\description{
The \code{plot_tt()} function allows for the insertion of images and inline plots into
tinytable objects. This function can handle both local and web-based images. #'
}
\details{
The \code{plot_tt()} can insert images and inline plots into tables.
}
\examples{
# HTML output can load images from the web, but not LaTeX
# See website for tutorials.
fn <- paste0(tempfile(), ".png")
dat <- data.frame("R" = "")
img <- "https://cran.r-project.org/Rlogo.svg"
tt(dat) |>
  plot_tt(i = 1, j = 1, path_img = img, height = 7) |>
  style_tt(j = 1, align = "c") |>
  save_tt(fn)

}
