% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tinytable}
\alias{print.tinytable}
\title{Print a tinytable to console or viewer pane}
\usage{
\method{print}{tinytable}(x, output = getOption("tinytable_print_output", default = NULL), ...)
}
\value{
launch a browser window or cat() the table to console.
}
\description{
Print a tinytable to console or viewer pane
}
