% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tt.R
\name{save_tt}
\alias{save_tt}
\title{Save a Tiny Table to File}
\usage{
save_tt(x, filename, overwrite = FALSE)
}
\arguments{
\item{x}{The tinytable object to be saved.}

\item{filename}{A string representing the path to the file where the object should be saved.}

\item{overwrite}{A logical value indicating whether to overwrite an existing file.}
}
\value{
invisible(TRUE)
}
\description{
This function saves an object of class tinytable to a specified file, with an option to overwrite existing files.
}
\examples{
\dontrun{

library(tinytable)
tab  <- tt(mtcars[1:4, 1:4])
save_tt(tt, "path/to/file.txt")

}

}
