% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_tt.R, R/zzz-aliases.R
\name{strip_tt}
\alias{strip_tt}
\alias{tt_strip}
\title{Strip elements from a Tiny Table}
\usage{
strip_tt(
  x,
  style = FALSE,
  format = FALSE,
  theme = FALSE,
  notes = FALSE,
  caption = FALSE,
  group = FALSE,
  bold = FALSE,
  italic = FALSE,
  monospace = FALSE,
  underline = FALSE,
  strikeout = FALSE,
  color = FALSE,
  background = FALSE,
  fontsize = FALSE,
  align = FALSE,
  alignv = FALSE,
  colspan = FALSE,
  rowspan = FALSE,
  indent = FALSE,
  line = FALSE,
  html_class = FALSE,
  html_css = FALSE,
  html_css_rule = FALSE,
  html_engine = FALSE,
  tabularray_inner = FALSE,
  tabularray_outer = FALSE,
  width = FALSE
)

tt_strip(
  x,
  style = FALSE,
  format = FALSE,
  theme = FALSE,
  notes = FALSE,
  caption = FALSE,
  group = FALSE,
  bold = FALSE,
  italic = FALSE,
  monospace = FALSE,
  underline = FALSE,
  strikeout = FALSE,
  color = FALSE,
  background = FALSE,
  fontsize = FALSE,
  align = FALSE,
  alignv = FALSE,
  colspan = FALSE,
  rowspan = FALSE,
  indent = FALSE,
  line = FALSE,
  html_class = FALSE,
  html_css = FALSE,
  html_css_rule = FALSE,
  html_engine = FALSE,
  tabularray_inner = FALSE,
  tabularray_outer = FALSE,
  width = FALSE
)
}
\arguments{
\item{x}{A table object created by \code{tt()}.}

\item{style}{\code{TRUE} reverts \code{style_tt()}.}

\item{format}{\code{TRUE} reverts \code{format_tt()}.}

\item{theme}{\code{TRUE} erases all themes and pre/post styling}

\item{notes}{\code{TRUE} reverts the effect of the \code{notes} argument from \code{tt()}.}

\item{caption}{\code{TRUE} reverts the effect of the \code{caption} argument from \code{tt()}.}

\item{group}{\code{TRUE} reverts \code{group_tt()}.}

\item{bold}{\code{TRUE} reverts the effect of the \code{bold} argument from \code{style_tt()}.}

\item{italic}{\code{TRUE} reverts the effect of the \code{italic} argument from \code{style_tt()}.}

\item{monospace}{\code{TRUE} reverts the effect of the \code{monospace} argument from \code{style_tt()}.}

\item{underline}{\code{TRUE} reverts the effect of the \code{underline} argument from \code{style_tt()}.}

\item{strikeout}{\code{TRUE} reverts the effect of the \code{strikeout} argument from \code{style_tt()}.}

\item{color}{\code{TRUE} reverts the effect of the \code{color} argument from \code{style_tt()}.}

\item{background}{\code{TRUE} reverts the effect of the \code{background} argument from \code{style_tt()}.}

\item{fontsize}{\code{TRUE} reverts the effect of the \code{fontsize} argument from \code{style_tt()}.}

\item{align}{\code{TRUE} reverts the effect of the \code{align} argument from \code{style_tt()}.}

\item{alignv}{\code{TRUE} reverts the effect of the \code{alignv} argument from \code{style_tt()}.}

\item{colspan}{\code{TRUE} reverts the effect of the \code{colspan} argument from \code{style_tt()}.}

\item{rowspan}{\code{TRUE} reverts the effect of the \code{rowspan} argument from \code{style_tt()}.}

\item{indent}{\code{TRUE} reverts the effect of the \code{indent} argument from \code{style_tt()}.}

\item{line}{\code{TRUE} reverts the effect of the \code{line} argument from \code{style_tt()}.}

\item{html_class}{\code{TRUE} reverts the effect of the \code{html_class} argument from \code{style_tt()}.}

\item{html_css}{\code{TRUE} reverts the effect of the \code{html_css} argument from \code{style_tt()}.}

\item{html_css_rule}{\code{TRUE} reverts the effect of the \code{html_css_rule} argument from \code{style_tt()}.}

\item{html_engine}{\code{TRUE} reverts the effect of the \code{engine} argument from \code{theme_html()}.}

\item{tabularray_inner}{\code{TRUE} reverts the effect of the \code{inner} argument from \code{theme_latex()}.}

\item{tabularray_outer}{\code{TRUE} reverts the effect of the \code{outer} argument from \code{theme_latex()}.}

\item{width}{\code{TRUE} reverts the effect of the \code{width} argument from \code{tt()}.}
}
\value{
An object of class \code{tt} representing the table with stripped styling.
}
\description{
Strip elements from a Tiny Table

Alias for \code{strip_tt()}
}
