% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_bootstrap.R
\name{style_bootstrap}
\alias{style_bootstrap}
\title{Style a Tiny Table for HTML Bootstrap output}
\usage{
style_bootstrap(x, i, j, css = NULL, css_rule = NULL)
}
\arguments{
\item{x}{A tiny table generated by the \code{tt()} function.}

\item{i}{An integer vector specifying the rows to be styled. If missing,
defaults to all rows in the table.}

\item{j}{An integer vector specifying the columns to be styled. If missing,
defaults to all columns in the table.}

\item{css}{A vector of CSS style declarations to be applied. Each element
corresponds to a cell defined by \code{i} and \code{j}.}

\item{css_rule}{A string with complete CSS rules that apply to the table class specified using the \code{theme} argument of the \code{tt()} function.}
}
\value{
Returns the modified HTML table object with added Bootstrap styling.
}
\description{
Applies Bootstrap styling to specific cells in an HTML table. It allows
customization of cell styles using CSS.
}
