% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_style.R
\name{style_eval,tinytable_grid-method}
\alias{style_eval,tinytable_grid-method}
\title{tinytable S4 method}
\usage{
\S4method{style_eval}{tinytable_grid}(x)
}
\description{
tinytable S4 method
}
\keyword{internal}
