% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_tt.R
\name{style_tt}
\alias{style_tt}
\title{export}
\usage{
style_tt(
  x,
  i,
  j,
  bold = FALSE,
  italic = FALSE,
  monospace = FALSE,
  color = NULL,
  background = NULL,
  fontsize = NULL,
  width = NULL,
  align = NULL,
  colspan = NULL,
  rowspan = NULL,
  tabularray = NULL,
  bootstrap = NULL
)
}
\description{
export
}
