% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_default.R
\name{theme_default}
\alias{theme_default}
\title{Default theme for TinyTable}
\usage{
theme_default(x, ...)
}
\arguments{
\item{x}{A tinytable object.}

\item{...}{Additional arguments are ignored.}
}
\value{
A modified \code{tinytable} object.
}
\description{
Default theme for TinyTable
}
