% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_empty.R
\name{theme_empty}
\alias{theme_empty}
\title{Theme for a void table}
\usage{
theme_empty(x, ...)
}
\arguments{
\item{x}{A tinytable object.}

\item{...}{Additional arguments are ignored.}
}
\description{
This function calls styles and formatting applied to a \code{tinytable} object up to that point in the pipeline.
}
