% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_latex.R
\name{theme_latex}
\alias{theme_latex}
\title{LaTeX-specific styles and options}
\usage{
theme_latex(x, inner = NULL, outer = NULL, environment = "tblr")
}
\arguments{
\item{x}{A \code{tinytable} object.}

\item{inner}{A string that specifies the "inner" settings of a tabularray LaTeX table.}

\item{outer}{A string that specifies the "outer" settings of a tabularray LaTeX table.}

\item{environment}{A string that specifies the LaTeX environment to use, default is "tblr".}
}
\description{
LaTeX-specific styles and options
}
