% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rotate.R
\name{theme_rotate}
\alias{theme_rotate}
\title{Rotate table theme (LaTeX and Typst only)}
\usage{
theme_rotate(x, angle = 90, ...)
}
\arguments{
\item{x}{A \code{tinytable} object.}

\item{angle}{Numeric. Rotation angle in degrees (0-360).}

\item{...}{Additional arguments (ignored).}
}
\value{
A modified \code{tinytable} object.
}
\description{
Rotate table theme (LaTeX and Typst only)
}
