% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tt.R
\name{theme_tt}
\alias{theme_tt}
\title{Themes for \code{tinytable}}
\usage{
theme_tt(x, theme, ...)
}
\arguments{
\item{x}{A \code{tinytable} object}

\item{theme}{String. Name of the theme to apply.
\itemize{
\item "resize" (LaTeX)
\item "multipage" (LaTeX)
}}

\item{...}{Additional arguments passed the themeing function. See the "Arguments" section below for a list of supported arguments for each theme.}
}
\value{
A modified \code{tinytable} object
}
\description{
A theme is a function that applies a collection of transformations to a \code{tinytable} object. Whereas the other functions in \code{tinytable} aim to be output-agnostic, themes can be output-specific, only applying to LaTeX or HTML, as needed. Themes can also have their own arguments
}
\section{Themes}{

\itemize{
\item resize: Resize a \code{tinytable} to fit the line width, while maintaining proportions
\item multipage: Long tables continue on the next page
}
}

\section{Arguments}{


resize
\itemize{
\item \code{width}: A numeric value between 0.01 and 1, representing the proportion of the line width to use
}

multipage
\itemize{
\item \code{rowhead}: Non-negative integer. The number of header rows to repeat on each page.
\item \code{rowfoot}: Non-negative integer. The number of footer rows to repeat on each page.
}
}

