% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tt.R
\docType{data}
\name{theme_tt}
\alias{theme_tt}
\title{Themes for tinytable}
\format{
An object of class \code{list} of length 1.
}
\usage{
theme_tt
}
\arguments{
\item{x}{A \code{tinytable} object}

\item{theme}{String. Name of the theme to apply.
\itemize{
\item "resize" (LaTeX)
\item "multipage" (LaTeX & Typst)
\item "scrollbox" (HTML)
\item "tooltip" (HTML)
}}
}
\description{
Themes for tinytable
}
\section{Themes}{

\itemize{
\item resize: Resize a \code{tinytable} to fit the line width, while maintaining proportions
}
}

\section{Arguments}{


resize
\itemize{
\item \code{width}: A numeric value between 0.01 and 1, representing the proportion of the line width to use
}

multipage
\itemize{
\item Blah blah blah
\item Blah blah blah
}

scrollbox
\itemize{
\item Blah blah blah
\item Blah blah blah
}

tooltip
\itemize{
\item Blah blah blah
\item Blah blah blah
}
}

\keyword{datasets}
