% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_typst.R
\name{theme_typst}
\alias{theme_typst}
\title{Typst-specific styles and options}
\usage{
theme_typst(
  x,
  figure = get_option("tinytable_typst_figure", default = TRUE),
  align_figure = get_option("tinytable_typst_align_figure", NULL),
  ...
)
}
\arguments{
\item{x}{A \code{tinytable} object.}

\item{figure}{Logical, whether to wrap the table in a Typst figure environment and block.}

\item{align_figure}{Character string indicating horizontal alignment: "l", "c", or "r".
Defaults to \code{get_option("tinytable_theme_placement_horizontal", NULL)}. When NULL, uses default center alignment.}

\item{...}{Additional arguments.}
}
\description{
Typst-specific styles and options
}
