% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_typst.R
\name{theme_typst}
\alias{theme_typst}
\title{Typst-specific styles and options}
\usage{
theme_typst(x, figure = TRUE, ...)
}
\arguments{
\item{x}{A \code{tinytable} object.}

\item{figure}{Logical, whether to wrap the table in a Typst figure environment and block.}

\item{...}{Additional arguments.}
}
\description{
Typst-specific styles and options
}
