% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_void.R
\name{theme_void}
\alias{theme_void}
\title{Theme for a void table}
\usage{
theme_void(x, ...)
}
\arguments{
\item{x}{A tinytable object.}

\item{...}{Additional arguments  are ignored.}
}
\description{
Theme for a void table
}
