\name{tinytable-package}
\alias{tinytable-package}
\alias{tinytable}
\docType{package}
\title{\packageTitle{tinytable}}
\description{\packageDescription{tinytable}}
\section{Package Content}{\packageIndices{tinytable}}
\author{\packageAuthor{tinytable}}
\section{Maintainer}{\packageMaintainer{tinytable}}
\keyword{package}
