% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\name{tt}
\alias{tt}
\title{Draw a table}
\usage{
tt(
  x,
  output = NULL,
  align = NULL,
  caption = NULL,
  theme = "default",
  extendable = FALSE
)
}
\arguments{
\item{x}{dataframe}

\item{output}{string "markdown", "latex", or "html". If \code{output} is \code{NULL}, then:
\itemize{
\item "html" if \code{knitr::is_html_output()} is \code{TRUE}
\item "latex" if \code{knitr::is_latex_output()} is \code{TRUE}
\item Otherwise determined by setting a global option: \code{options(tt_output_default = "markdown")}
}}

\item{theme}{string
\itemize{
\item LaTeX: "default", "striped", "void", or "grid".
\item HTML: "default", "striped", "void", "grid", or a (composite) Bootstrap class such as \code{"table table-dark"} or \code{"table table-dark table-hover"}. See
}}
}
\description{
Draw a table
}
