% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt.R
\name{tt}
\alias{tt}
\title{Draw a Tiny Table}
\usage{
tt(
  x,
  output = NULL,
  align = NULL,
  caption = NULL,
  width = NULL,
  notes = NULL,
  theme = "default",
  placement = getOption("tt_tabularray_placement", default = NULL)
)
}
\arguments{
\item{x}{A data frame or data table to be rendered as a table.}

\item{output}{The format of the output table. Can be "html", "latex", or "markdown". If NULL, the format is automatically detected in Quarto or Rmarkdown documents.}

\item{align}{A string specifying the alignment of columns. Each character in the string corresponds to a column; 'l' for left, 'c' for center, and 'r' for right alignment. The length of the string must match the number of columns in \code{x}.}

\item{caption}{A string that will be used as the caption of the table.}

\item{width}{A numeric value between 0 and 1 indicating the proportion of the line width that the table should cover.}

\item{notes}{A single string or a (named) list of strings to append at the bottom of the table.}

\item{theme}{The theme to apply to the table.
\itemize{
\item LaTeX: "default", "striped", "void", or "grid".
\item HTML: "default", "striped", "void", "grid", or a (composite) Bootstrap class such as \code{"table table-dark"} or \code{"table table-dark table-hover"}. See
}}

\item{placement}{A string to control the position of tables in LaTeX. Will be inserted in square brackets like: \verb{\\\\begin\{table\}[H]}}
}
\value{
An object of class \code{tt} representing the table.
}
\description{
The \code{tt} function renders a table in different formats (HTML, Markdown, or LaTeX) with various styling options.
}
\section{LaTeX preamble}{


To be able to use all features of \code{tinytable} in PDF (LaTeX) documents, these commands need to be placed in your LaTeX preamble:

\if{html}{\out{<div class="sourceCode latex">}}\preformatted{\\usepackage\{float\}
\\usepackage\{codehigh\}
\\usepackage\{tabularray\}
\\UseTblrLibrary\{booktabs\}
\\NewTableCommand\{\\tinytableDefineColor\}[3]\{\\definecolor\{#1\}\{#2\}\{#3\}\}
}\if{html}{\out{</div>}}
}

\examples{
\dontshow{if (getOption("tt_local", default = FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
library(tinytable)
x <- mtcars[1:4, 1:5]

tt(x)

tt(x,
  theme = "striped",
  width = 0.5,
  align = "ccrrl",
  caption = "Data about cars.")

}
